requirements.txt
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
idna==2.8
ipython-genutils==0.2.0
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-core==4.4.0
nbformat==4.4.0
numpy==1.16.3
plotly==3.9.0
pyrsistent==0.15.2
pytz==2019.1
requests==2.22.0
retrying==1.3.3
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
pyyaml
prometheus_client
